# openapi_patched.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/users/me": {
      "put": {
        "summary": "更新用户信息",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要持有合法 Token\n【功能描述】编辑和更新当前登录账号的公开基本信息（如昵称、个人签名、头像和联系方式等）",
        "operationId": "更新用户信息",
        "tags": [
          "用户模块"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              },
              "example": {
                "avatar": "https://example.com/new-avatar.png",
                "email": "new.email@example.com",
                "nickname": "新昵称"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMeResponse"
                },
                "example": {
                  "code": 200,
                  "message": "资料已更新"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "409": {
            "description": "资源冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 409,
                  "message": "资源已存在或状态冲突"
                }
              }
            },
            "headers": {},
            "x-apifox-ordering": 5
          },
          "422": {
            "description": "实体无法处理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 422,
                  "message": "请求格式正确但语义错误",
                  "details": {
                    "field": "email",
                    "reason": "格式不正确"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "参数错误",
            "x-apifox-ordering": 6
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 7
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 8
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 9
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "用户模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441885895-run"
      },
      "get": {
        "summary": "获取当前用户信息",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【前置条件】需要持有合法 Token\n【功能描述】解析用户当前身份，获取自己账号的详细数据和权限配置。",
        "operationId": "获取当前用户信息",
        "tags": [
          "用户模块"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeListResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "avatar": "https://example.com/avatar.png",
                    "email": "admin@example.com",
                    "id": 10001,
                    "nickname": "管理员",
                    "phone": "13800138000",
                    "status": "active",
                    "username": "admin01"
                  },
                  "message": "成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 5
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 6
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 7
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-apifox-folder": "用户模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441707422-run"
      }
    },
    "/api/v1/users/register": {
      "post": {
        "summary": "用户注册",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【功能描述】新用户使用账号密码等资料注册加入平台",
        "operationId": "用户注册",
        "tags": [
          "用户模块"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegisterRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "StrongPassword!2",
                "username": "new_user_88"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRegisterResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "id": 10025,
                    "username": "new_user_88"
                  },
                  "message": "注册成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "409": {
            "description": "资源冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 409,
                  "message": "资源已存在或状态冲突"
                }
              }
            },
            "headers": {},
            "x-apifox-ordering": 5
          },
          "422": {
            "description": "实体无法处理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 422,
                  "message": "请求格式正确但语义错误",
                  "details": {
                    "field": "email",
                    "reason": "格式不正确"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "参数错误",
            "x-apifox-ordering": 6
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 7
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 8
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 9
          }
        },
        "security": [],
        "x-apifox-folder": "用户模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441703036-run"
      }
    },
    "/api/v1/users/login": {
      "post": {
        "summary": "用户登录",
        "deprecated": false,
        "description": "【版本】v1\n【环境】REST 接口\n【功能描述】用户名密码或手机号密码等多方式登录系统完成认证",
        "operationId": "用户登录",
        "tags": [
          "用户模块"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoginRequest"
              },
              "example": {
                "account": "user123",
                "password": "Password123!"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLoginResponse"
                },
                "example": {
                  "code": 200,
                  "data": {
                    "accessToken": "eyJh...xxxx",
                    "expiresIn": 7200,
                    "refreshToken": "refx...xxxx"
                  },
                  "message": "登录成功"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功",
            "x-apifox-ordering": 0
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 400,
                  "message": "请求参数错误",
                  "details": {
                    "field": "name",
                    "reason": "不能为空"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "请求有误",
            "x-apifox-ordering": 1
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "message": "未授权，请先登录"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "没有权限",
            "x-apifox-ordering": 2
          },
          "403": {
            "description": "禁止访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 403,
                  "message": "无权限访问此资源"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "禁止访问",
            "x-apifox-ordering": 3
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 404,
                  "message": "请求的资源不存在"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "记录不存在",
            "x-apifox-ordering": 4
          },
          "409": {
            "description": "资源冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 409,
                  "message": "资源已存在或状态冲突"
                }
              }
            },
            "headers": {},
            "x-apifox-ordering": 5
          },
          "422": {
            "description": "实体无法处理",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 422,
                  "message": "请求格式正确但语义错误",
                  "details": {
                    "field": "email",
                    "reason": "格式不正确"
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "参数错误",
            "x-apifox-ordering": 6
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 500,
                  "message": "服务器内部错误，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器错误",
            "x-apifox-ordering": 7
          },
          "502": {
            "description": "网关错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 502,
                  "message": "网关错误，上游服务不可用"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "网关错误",
            "x-apifox-ordering": 8
          },
          "503": {
            "description": "服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 503,
                  "message": "服务暂时不可用，请稍后重试"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "服务器故障",
            "x-apifox-ordering": 9
          }
        },
        "security": [],
        "x-apifox-folder": "用户模块",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/8090070/apis/api-441697162-run"
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "标准错误响应",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "details": {
            "type": "object",
            "description": "详细信息",
            "x-apifox-orders": [],
            "properties": {},
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "code",
          "message"
        ],
        "x-apifox-orders": [
          "code",
          "message",
          "details"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "UpdateMeRequest": {
        "properties": {
          "avatar": {
            "description": "新头像链接",
            "type": "string"
          },
          "email": {
            "description": "更新电子邮箱",
            "type": "string"
          },
          "nickname": {
            "description": "新昵称",
            "type": "string"
          },
          "phone": {
            "description": "更新手机号码",
            "type": "string"
          }
        },
        "type": "object",
        "x-apifox-orders": [
          "avatar",
          "email",
          "nickname",
          "phone"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "CreateLoginRequest": {
        "properties": {
          "account": {
            "description": "用户名、手机号或邮箱",
            "type": "string"
          },
          "password": {
            "description": "登录密码",
            "type": "string"
          }
        },
        "required": [
          "account",
          "password"
        ],
        "type": "object",
        "x-apifox-orders": [
          "account",
          "password"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "CreateRegisterRequest": {
        "properties": {
          "email": {
            "description": "绑定邮箱（可选）",
            "type": "string"
          },
          "password": {
            "description": "设置登录密码",
            "type": "string"
          },
          "username": {
            "description": "设置登录用户名",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "x-apifox-orders": [
          "email",
          "password",
          "username"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "MeListResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "data": {
            "description": "用户详情数据",
            "properties": {
              "avatar": {
                "description": "头像缩略图链接",
                "type": "string"
              },
              "email": {
                "description": "电子邮箱",
                "type": "string"
              },
              "id": {
                "description": "真实用户ID",
                "type": "integer"
              },
              "nickname": {
                "description": "页面展示昵称",
                "type": "string"
              },
              "phone": {
                "description": "联系手机",
                "type": "string"
              },
              "status": {
                "description": "账户状态：active/disabled",
                "type": "string"
              },
              "username": {
                "description": "账户登录名",
                "type": "string"
              }
            },
            "type": "object",
            "x-apifox-orders": [
              "avatar",
              "email",
              "id",
              "nickname",
              "phone",
              "status",
              "username"
            ],
            "x-apifox-ignore-properties": []
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "data",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "UpdateMeResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "CreateLoginResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "data": {
            "description": "认证数据",
            "properties": {
              "accessToken": {
                "description": "访问令牌（Bearer Token）",
                "type": "string"
              },
              "expiresIn": {
                "description": "过期时间阈值（秒）",
                "type": "integer"
              },
              "refreshToken": {
                "description": "用于换取新令牌的刷新令牌",
                "type": "string"
              }
            },
            "type": "object",
            "x-apifox-orders": [
              "accessToken",
              "expiresIn",
              "refreshToken"
            ],
            "x-apifox-ignore-properties": []
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "data",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "CreateRegisterResponse": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "data": {
            "description": "注册新建用户基本数据",
            "properties": {
              "id": {
                "description": "用户ID",
                "type": "integer"
              },
              "username": {
                "description": "用户名",
                "type": "string"
              }
            },
            "type": "object",
            "x-apifox-orders": [
              "id",
              "username"
            ],
            "x-apifox-ignore-properties": []
          },
          "message": {
            "description": "提示信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-apifox-orders": [
          "code",
          "data",
          "message"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}